pyproject.toml
[project]
name = "code-analyzer"
version = "0.1.0"
description = "AI-powered code analyzer agent that analyzes code and generates structured reports"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "pydantic-ai>=0.0.30",
    "aiofiles>=24.0.0",
    "docker>=7.0.0",
    "httpx>=0.28.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
